Services write explictely to access log
Previously, writing to access log was happening within
on-close handler for the channel. The problem was that
the on-close handler does not have enough information
about response code and error message. Now the implementation
is modified so that every service explicetely writes to
access log.

Additionally, all unit tests verify that the correct status
code and error message are written to access log, and that
there is exactly one record in access log per one invocation of
an API service.